requirements.txt
rdkit
numpy===1.26.4
python-dotenv===1.0.1